init/requirements.txt
mysql-connector
requests